perf(rust)!: small-integer token representation
Replace Value::Int(BigInt) with Value::Int(Integer), where Integer stores
i64-range values inline and only falls back to BigInt for larger ones (the
variants are hidden behind a struct so the "big only when out of range"
invariant holds, keeping Eq/Ord exact). In-range integers — the common
case — no longer heap-allocate. ascii_char, utf8_char, integer, line, and
byte_offset emit Integer; the codegen macros use a parse_integer helper.

Datetime benchmark: interpreter ~1.16 -> ~0.81 us, specialized
compile_parser! ~0.56 -> ~0.14 us, fair token ceiling ~0.43 -> ~0.07 us.

+++ rust/tests/choice_and_repeat.rs
@@ -1,10 +1,10 @@
+use nimble_parsec_rs::Integer;
 use nimble_parsec_rs::{
     ascii_char, choice, concat, optional, repeat, string, AsciiPredicate, Value,
 };
-use num_bigint::BigInt;
 
 fn ch(c: char) -> Value {
-    Value::Int(BigInt::from(c as u32))
+    Value::Int(Integer::from(c as u32))
 }
 
 #[test]
